Revert "added footer"
