--- app/components/Breadcrumbs.tsx
@@ -1,84 +1,0 @@
-import Link from 'next/link'
-
-interface BreadcrumbItem {
-  name: string
-  href: string
-  current?: boolean
-}
-
-interface BreadcrumbsProps {
-  items: BreadcrumbItem[]
-  className?: string
-}
-
-export default function Breadcrumbs({ items, className = '' }: BreadcrumbsProps) {
-  const jsonLd = {
-    '@context': 'https://schema.org',
-    '@type': 'BreadcrumbList',
-    'itemListElement': items.map((item, index) => ({
-      '@type': 'ListItem',
-      'position': index + 1,
-      'name': item.name,
-      'item': `https://onetimeemail.net${item.href}`
-    }))
-  }
-
-  return (
-    <>
-      {/* Structured Data for Breadcrumbs */}
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
-      />
-      
-      <nav className={`flex ${className}`} aria-label="Breadcrumb">
-        <ol className="inline-flex items-center space-x-1 md:space-x-3">
-          {items.map((item, index) => (
-            <li key={item.href} className="inline-flex items-center">
-              {index > 0 && (
-                <svg
-                  className="w-3 h-3 text-gray-400 mx-1"
-                  aria-hidden="true"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 6 10"
-                >
-                  <path
-                    stroke="currentColor"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="m1 9 4-4-4-4"
-                  />
-                </svg>
-              )}
-              {item.current ? (
-                <span className="text-sm font-medium text-gray-500 cursor-default">
-                  {item.name}
-                </span>
-              ) : (
-                <Link
-                  href={item.href}
-                  className="inline-flex items-center text-sm font-medium text-gray-700 hover:text-blue-600 transition-colors"
-                >
-                  {index === 0 && (
-                    <svg
-                      className="w-3 h-3 mr-2.5"
-                      aria-hidden="true"
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                    >
-                      <path d="m19.707 9.293-2-2-7-7a1 1 0 0 0-1.414 0l-7 7-2 2a1 1 0 0 0 1.414 1.414L2 10.414V18a2 2 0 0 0 2 2h3a1 1 0 0 0 1-1v-4a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v4a1 1 0 0 0 1 1h3a2 2 0 0 0 2-2v-7.586l.293.293a1 1 0 0 0 1.414-1.414Z" />
-                    </svg>
-                  )}
-                  {item.name}
-                </Link>
-              )}
-            </li>
-          ))}
-        </ol>
-      </nav>
-    </>
-  )
-}
