--- app/faq/page.tsx
@@ -1,331 +1,0 @@
-import { Metadata, Viewport } from 'next'
-import Link from 'next/link'
-import { generateAdvancedMetadata, generateAdvancedViewport } from '../components/AdvancedSEO'
-
-export const metadata: Metadata = generateAdvancedMetadata({
-  title: 'FAQ - Frequently Asked Questions about OneTimeEmail',
-  description: 'Get answers to common questions about OneTimeEmail temporary email service. Learn how to use disposable emails safely and protect your privacy.',
-  canonical: 'https://onetimeemail.net/faq'
-})
-
-export const viewport: Viewport = generateAdvancedViewport()
-
-// FAQ Structured Data for Rich Snippets
-const faqStructuredData = {
-  "@context": "https://schema.org",
-  "@type": "FAQPage",
-  "mainEntity": [
-    {
-      "@type": "Question",
-      "name": "How long do temporary emails last?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "Temporary emails generated by OneTimeEmail last for 10 minutes by default. After this time, the email address and all received messages are automatically deleted for your privacy and security."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "Is OneTimeEmail free to use?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "Yes, OneTimeEmail is completely free to use. There are no registration requirements, no hidden fees, no premium plans, and no subscription costs. Our service will always remain free."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "Do I need to register or create an account?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "No registration or account creation is required. Simply visit our website and click 'Generate Your Inbox' to get a temporary email address instantly. This ensures complete anonymity."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "Can I extend the expiration time of my temporary email?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "Currently, all temporary emails expire after 10 minutes for security and privacy reasons. This ensures your data is not stored longer than necessary and maintains optimal privacy protection."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "Is my privacy protected when using OneTimeEmail?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "Yes, OneTimeEmail prioritizes your privacy. We don't track users, store personal information, require registration, or use tracking cookies. All emails are automatically deleted after 10 minutes."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "What can I use temporary emails for?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "Temporary emails are perfect for online registrations, free trials, software downloads, newsletter signups, testing applications, avoiding spam, and any situation where you don't want to use your real email address."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "Are there any limitations on email types I can receive?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "You can receive most types of emails including verification emails, newsletters, and notifications. We block spam and malicious content for security. Emails larger than 1MB may be rejected for performance reasons."
-      }
-    },
-    {
-      "@type": "Question",
-      "name": "How secure is OneTimeEmail?",
-      "acceptedAnswer": {
-        "@type": "Answer",
-        "text": "OneTimeEmail uses HTTPS encryption, secure cloud hosting, automatic data deletion, and follows security best practices. We employ multiple layers of protection including rate limiting and content filtering."
-      }
-    }
-  ]
-}
-
-interface FAQItemProps {
-  question: string
-  answer: string
-  index: number
-}
-
-function FAQItem({ question, answer, index }: FAQItemProps) {
-  return (
-    <div className="bg-white/80 backdrop-blur-xl rounded-2xl p-8 border border-gray-200 shadow-xl">
-      <h3 className="text-xl font-semibold text-gray-900 mb-4 flex items-start gap-3">
-        <span className="bg-blue-500 text-white rounded-full w-8 h-8 flex items-center justify-center text-sm font-bold shrink-0 mt-1">
-          {index + 1}
-        </span>
-        {question}
-      </h3>
-      <div className="pl-11">
-        <p className="text-gray-700 leading-relaxed">{answer}</p>
-      </div>
-    </div>
-  )
-}
-
-export default function FAQPage() {
-  const faqs = [
-    {
-      question: "How long do temporary emails last?",
-      answer: "Temporary emails generated by OneTimeEmail last for 10 minutes by default. After this time, the email address and all received messages are automatically deleted for your privacy and security. This duration is designed to give you enough time to receive verification emails or sign up for services while ensuring your data doesn't linger unnecessarily."
-    },
-    {
-      question: "Is OneTimeEmail free to use?",
-      answer: "Yes, OneTimeEmail is completely free to use. There are no registration requirements, no hidden fees, no premium plans, and no subscription costs. Our service will always remain free as we believe privacy protection should be accessible to everyone without financial barriers."
-    },
-    {
-      question: "Do I need to register or create an account?",
-      answer: "No registration or account creation is required. Simply visit our website and click 'Generate Your Inbox' to get a temporary email address instantly. This approach ensures complete anonymity and protects your privacy by not collecting any personal information."
-    },
-    {
-      question: "Can I extend the expiration time of my temporary email?",
-      answer: "Currently, all temporary emails expire after 10 minutes for security and privacy reasons. This fixed duration ensures your data is not stored longer than necessary and maintains optimal privacy protection. If you need a longer-lasting email solution, consider using multiple temporary emails in sequence."
-    },
-    {
-      question: "Is my privacy protected when using OneTimeEmail?",
-      answer: "Yes, OneTimeEmail prioritizes your privacy above all else. We don't track users, store personal information, require registration, or use tracking cookies. All emails are automatically deleted after 10 minutes, and we employ zero-knowledge architecture where even our administrators cannot access your email content."
-    },
-    {
-      question: "What can I use temporary emails for?",
-      answer: "Temporary emails are perfect for online registrations, free trials, software downloads, newsletter signups, testing applications, avoiding spam, accessing gated content, one-time verifications, and any situation where you don't want to use your real email address. They're ideal for maintaining privacy while accessing digital services."
-    },
-    {
-      question: "Are there any limitations on email types I can receive?",
-      answer: "You can receive most types of emails including verification emails, newsletters, notifications, and confirmations. We automatically block spam and malicious content for security. Emails larger than 1MB may be rejected for performance reasons, and we filter out suspicious content to protect users."
-    },
-    {
-      question: "How secure is OneTimeEmail?",
-      answer: "OneTimeEmail uses enterprise-grade security including HTTPS/TLS encryption, secure cloud hosting on Vercel, automatic data deletion, and follows security best practices. We employ multiple layers of protection including rate limiting, content filtering, HMAC signature verification, and automated vulnerability monitoring."
-    },
-    {
-      question: "Can I receive emails with attachments?",
-      answer: "Yes, you can receive emails with attachments, but there are size limitations for performance and security reasons. Attachments are automatically scanned for malware, and the total email size (including attachments) cannot exceed 1MB. Attachments are deleted along with the email after 10 minutes."
-    },
-    {
-      question: "How do I know if I've received an email?",
-      answer: "When you generate a temporary email, you'll be taken to an inbox page that automatically refreshes to show new emails. You'll see emails appear in real-time as they arrive. The inbox displays the sender, subject, and timestamp of each received email."
-    },
-    {
-      question: "What happens if someone sends an email to an expired address?",
-      answer: "If someone sends an email to an expired temporary address, they will receive an automated bounce notification explaining that the mailbox does not exist or has expired. This helps senders understand that their email wasn't delivered and maintains the temporary nature of our service."
-    },
-    {
-      question: "Can I create multiple temporary emails at once?",
-      answer: "You can generate new temporary emails whenever you need them, but each email address is independent with its own 10-minute lifespan. If you need multiple addresses simultaneously, you can open multiple browser tabs or windows, each with its own temporary email."
-    },
-    {
-      question: "Is OneTimeEmail available in different languages?",
-      answer: "Currently, OneTimeEmail is available in English. However, you can receive emails in any language - our service processes emails in all character sets and languages. The interface language support may be expanded in future updates based on user demand."
-    },
-    {
-      question: "What makes OneTimeEmail different from other temporary email services?",
-      answer: "OneTimeEmail focuses on privacy, security, and simplicity. We offer transparent privacy practices, comprehensive security measures, no tracking or data collection, automatic HTTPS encryption, reliable email delivery, and a clean, user-friendly interface. Our commitment to keeping the service completely free also sets us apart."
-    },
-    {
-      question: "Can I use OneTimeEmail for business purposes?",
-      answer: "Yes, OneTimeEmail can be used for legitimate business purposes such as testing email functionalities, creating demo accounts, accessing research materials, or signing up for business tools and services. However, ensure compliance with the terms of service of the platforms you're registering with."
-    },
-    {
-      question: "What should I do if I'm not receiving emails?",
-      answer: "If you're not receiving expected emails, check that the sender has the correct temporary email address, ensure the email isn't blocked by size or content filters, verify that the 10-minute window hasn't expired, and check if the sender's email service is working properly. You can also try generating a new temporary email if needed."
-    }
-  ]
-
-  return (
-    <>
-      {/* Structured Data for Rich Snippets */}
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(faqStructuredData) }}
-      />
-      
-      <div className="min-h-screen bg-gradient-to-br from-white via-gray-50 to-blue-50">
-        <div className="container mx-auto px-4 py-12 max-w-4xl">
-          <header className="text-center mb-16">
-            <h1 className="text-5xl font-bold text-gray-900 mb-6">
-              Frequently Asked Questions
-            </h1>
-            <p className="text-xl text-gray-600 leading-relaxed max-w-3xl mx-auto">
-              Get answers to common questions about OneTimeEmail and learn how to make the most 
-              of our temporary email service while protecting your privacy.
-            </p>
-          </header>
-
-          <main>
-            {/* Quick Start Section */}
-            <section className="mb-16">
-              <div className="bg-gradient-to-r from-blue-500 to-cyan-500 rounded-2xl p-8 text-center text-white mb-12">
-                <h2 className="text-3xl font-bold mb-4">New to OneTimeEmail?</h2>
-                <p className="text-xl mb-6 opacity-90">
-                  Get started in seconds with our simple 3-step process
-                </p>
-                <div className="grid grid-cols-1 md:grid-cols-3 gap-6 text-center">
-                  <div>
-                    <div className="text-4xl mb-2">1️⃣</div>
-                    <h3 className="font-semibold mb-2">Generate</h3>
-                    <p className="text-sm opacity-80">Click to create your temporary email</p>
-                  </div>
-                  <div>
-                    <div className="text-4xl mb-2">2️⃣</div>
-                    <h3 className="font-semibold mb-2">Use</h3>
-                    <p className="text-sm opacity-80">Register or sign up with the temporary address</p>
-                  </div>
-                  <div>
-                    <div className="text-4xl mb-2">3️⃣</div>
-                    <h3 className="font-semibold mb-2">Receive</h3>
-                    <p className="text-sm opacity-80">Check your inbox for incoming emails</p>
-                  </div>
-                </div>
-                <Link 
-                  href="/"
-                  className="inline-block bg-white text-blue-600 px-8 py-4 rounded-xl font-semibold text-lg hover:bg-gray-100 transition-colors mt-6"
-                >
-                  Try It Now - Free
-                </Link>
-              </div>
-            </section>
-
-            {/* FAQ Items */}
-            <section className="mb-16">
-              <h2 className="text-3xl font-bold text-gray-900 mb-12 text-center">
-                Common Questions & Answers
-              </h2>
-              <div className="space-y-8">
-                {faqs.map((faq, index) => (
-                  <FAQItem
-                    key={index}
-                    question={faq.question}
-                    answer={faq.answer}
-                    index={index}
-                  />
-                ))}
-              </div>
-            </section>
-
-            {/* Additional Help Section */}
-            <section className="mb-16">
-              <div className="bg-white/80 backdrop-blur-xl rounded-2xl p-8 border border-gray-200 shadow-xl text-center">
-                <h2 className="text-2xl font-bold text-gray-900 mb-4">
-                  Still Have Questions?
-                </h2>
-                <p className="text-gray-700 mb-6">
-                  Can't find what you're looking for? Check out our other helpful resources:
-                </p>
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-                  <Link 
-                    href="/how-it-works"
-                    className="bg-blue-50 hover:bg-blue-100 p-4 rounded-lg transition-colors"
-                  >
-                    <div className="text-2xl mb-2">📖</div>
-                    <h3 className="font-semibold text-gray-900 mb-1">How It Works</h3>
-                    <p className="text-sm text-gray-600">Step-by-step guide</p>
-                  </Link>
-                  <Link 
-                    href="/about"
-                    className="bg-green-50 hover:bg-green-100 p-4 rounded-lg transition-colors"
-                  >
-                    <div className="text-2xl mb-2">🏢</div>
-                    <h3 className="font-semibold text-gray-900 mb-1">About Us</h3>
-                    <p className="text-sm text-gray-600">Our mission & values</p>
-                  </Link>
-                  <Link 
-                    href="/privacy"
-                    className="bg-purple-50 hover:bg-purple-100 p-4 rounded-lg transition-colors"
-                  >
-                    <div className="text-2xl mb-2">🔒</div>
-                    <h3 className="font-semibold text-gray-900 mb-1">Privacy Policy</h3>
-                    <p className="text-sm text-gray-600">How we protect you</p>
-                  </Link>
-                  <Link 
-                    href="/terms"
-                    className="bg-orange-50 hover:bg-orange-100 p-4 rounded-lg transition-colors"
-                  >
-                    <div className="text-2xl mb-2">📋</div>
-                    <h3 className="font-semibold text-gray-900 mb-1">Terms</h3>
-                    <p className="text-sm text-gray-600">Service terms</p>
-                  </Link>
-                </div>
-              </div>
-            </section>
-
-            {/* Best Practices Section */}
-            <section className="mb-16">
-              <h2 className="text-3xl font-bold text-gray-900 mb-8 text-center">
-                Best Practices for Temporary Emails
-              </h2>
-              <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-                <div className="bg-white/80 backdrop-blur-xl rounded-2xl p-8 border border-gray-200 shadow-xl">
-                  <h3 className="text-xl font-semibold text-gray-900 mb-4 flex items-center gap-3">
-                    <span>✅</span> Do
-                  </h3>
-                  <ul className="space-y-3 text-gray-700">
-                    <li>• Use for one-time registrations and verifications</li>
-                    <li>• Copy important information before expiration</li>
-                    <li>• Use for testing websites and applications</li>
-                    <li>• Generate new emails for different services</li>
-                    <li>• Keep the inbox tab open while waiting for emails</li>
-                  </ul>
-                </div>
-                <div className="bg-white/80 backdrop-blur-xl rounded-2xl p-8 border border-gray-200 shadow-xl">
-                  <h3 className="text-xl font-semibold text-gray-900 mb-4 flex items-center gap-3">
-                    <span>❌</span> Don't
-                  </h3>
-                  <ul className="space-y-3 text-gray-700">
-                    <li>• Use for important long-term accounts</li>
-                    <li>• Share temporary email addresses publicly</li>
-                    <li>• Rely on temporary emails for password resets</li>
-                    <li>• Use for financial or sensitive accounts</li>
-                    <li>• Expect emails to remain accessible after expiration</li>
-                  </ul>
-                </div>
-              </div>
-            </section>
-          </main>
-
-
-        </div>
-      </div>
-    </>
-  )
-}
