--- app/robots.ts
@@ -1,25 +1,0 @@
-import { MetadataRoute } from 'next'
-
-export default function robots(): MetadataRoute.Robots {
-  return {
-    rules: [
-      {
-        userAgent: '*',
-        allow: '/',
-        disallow: ['/api/', '/admin/', '/inbox/*'],
-      },
-      {
-        userAgent: 'Googlebot',
-        allow: '/',
-        disallow: ['/api/', '/admin/', '/inbox/*'],
-      },
-      {
-        userAgent: 'Bingbot',
-        allow: '/',
-        disallow: ['/api/', '/admin/', '/inbox/*'],
-      },
-    ],
-    sitemap: 'https://onetimeemail.net/sitemap.xml',
-    host: 'https://onetimeemail.net',
-  }
-}
